# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shirone Backend API
  version: 0.1.0
  description: Provider-neutral API for Shirone dynamic edition.
paths:
  /api/v1/media/{key}:
    get:
      operationId: getMedia
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Immutable media object.
        '404':
          description: Media was not found.
  /api/v1/admin/media/:
    get:
      operationId: listMedia
      responses:
        '200':
          description: Media metadata list.
    post:
      operationId: uploadMedia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                altText:
                  type: string
      responses:
        '201':
          description: Media object stored and metadata created.
        '413':
          description: Upload exceeds the configured size limit.
        '415':
          description: Media type is not supported.
  /api/v1/taxonomies:
    get:
      operationId: listTaxonomies
      responses:
        '200':
          description: Taxonomies and their terms.
  /api/v1/admin/taxonomies/:
    post:
      operationId: createTaxonomy
      responses:
        '201':
          description: Taxonomy created by an editor or administrator.
  /api/v1/admin/taxonomies/{id}/terms:
    post:
      operationId: createTaxonomyTerm
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Term created in the taxonomy.
  /api/v1/admin/users/:
    get:
      operationId: listUsers
      responses:
        '200':
          description: User list visible to administrators.
        '403':
          description: Administrator access is required.
  /api/v1/admin/users/{id}:
    put:
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User role and status updated.
        '409':
          description: The last active administrator cannot be changed.
  /api/v1/content/{slug}/comments:
    get:
      operationId: listPublishedComments
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved comments for published content.
    post:
      operationId: createComment
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Pending comment created by an authenticated user.
  /api/v1/admin/comments/:
    get:
      operationId: listManagedComments
      responses:
        '200':
          description: Comments visible to editors and administrators.
  /api/v1/admin/comments/{id}:
    put:
      operationId: moderateComment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comment moderation status updated.
  /api/v1/content/:
    get:
      operationId: listPublishedContent
      responses:
        '200':
          description: Published content list.
  /api/v1/content/{slug}:
    get:
      operationId: getPublishedContent
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Published content item.
        '404':
          description: Content was not found.
  /api/v1/admin/content/:
    get:
      operationId: listManagedContent
      responses:
        '200':
          description: Content list including drafts and archived items.
    post:
      operationId: createContent
      responses:
        '201':
          description: Content created by an editor or administrator.
        '401':
          description: Authentication is required.
        '403':
          description: Editor access is required.
  /api/v1/admin/content/{id}:
    get:
      operationId: getManagedContent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Managed content item.
    put:
      operationId: updateContent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Content updated by an editor or administrator.
    delete:
      operationId: archiveContent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Content archived and revision recorded.
  /api/v1/admin/content/{id}/revisions:
    get:
      operationId: listContentRevisions
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Immutable content revision snapshots.
  /api/v1/auth/register:
    post:
      operationId: register
      responses:
        '201':
          description: Account and session created.
        '409':
          description: Email or username already exists.
  /api/v1/auth/login:
    post:
      operationId: login
      responses:
        '200':
          description: Session created.
        '401':
          description: Invalid credentials.
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses:
        '204':
          description: Session removed.
  /api/v1/auth/me:
    get:
      operationId: currentUser
      responses:
        '200':
          description: Current account.
        '401':
          description: Authentication is required.
  /api/v1/health:
    get:
      operationId: health
      responses:
        '200':
          description: Process is alive.
  /api/v1/ready:
    get:
      operationId: readiness
      responses:
        '200':
          description: Database is reachable.
        '503':
          description: Database is unavailable.